feat: Add ControllsComponent with color input
This commit adds a new component called ControllsComponent, which is responsible for rendering a color input element. The color input allows users to select a color by clicking on it. This component is useful for applications that require user input for color selection.

The ControllsComponent is implemented as a custom HTML element using the Web Components API. It extends the HTMLElement class and uses the shadow DOM to encapsulate its internal implementation. The color input element is added to the shadow DOM using the attachShadow() method.

The color input element has an id of "outsideColor" and a default value of "#ffffff". This default value represents the color white. Developers can access the selected color value using JavaScript.

This commit does not include any other changes or modifications.